+++ lightweight/__init__.py
@@ -1,5 +1,6 @@
-from .content import markdown, render, template
-from .files import paths
 from .site import Site
+from .content import markdown, render
+from .files import paths
+from .template import template
 
 __version__ = '0.1.0.dev9'
